Phew. Pushing through.

+++ Core/Utility.h
@@ -42,8 +42,10 @@ T Reverse(T * var)
 #ifdef _CRT_USE_BUILTIN_OFFSETOF
 #define offsetof(s, m) __builtin_offsetof(s, m)
 #else
-#define offsetof(s, m) reinterpret_cast<uint64>(&(*reinterpret_cast<s *>(0)).m)
+#define offsetof(s, m) reinterpret_cast<uint32>(&(*reinterpret_cast<s *>(0)).m)
 #endif
+
+// @Research: Why uint64 again?
 
 template <uint8 count> struct GetDataTypeByValueHelper {};
